requirements.txt
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
openpyxl==3.1.2
#picassosr==
moviepy==1.0.3
python-dotenv==1.0.1
pre-commit==3.5.1
bandit==1.7.8
coverage==7.4.4
isort==5.13.2
